requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
dash>=2.14.0
scipy>=1.11.0
openai>=1.0.0
python-dotenv>=1.0.0

